+++ DiscordBot/OpenShockDiscordDb/UsersFriendwhitelist.cs
@@ -5,9 +5,9 @@ namespace OpenShock.DiscordBot.OpenShockDiscordDb;
 
 public partial class UsersFriendwhitelist
 {
-    public decimal User { get; set; }
+    public ulong User { get; set; }
 
-    public decimal WhitelistedFriend { get; set; }
+    public ulong WhitelistedFriend { get; set; }
 
     public DateTime CreatedOn { get; set; }
